Update to use Fastlane watch and improve docs

+++ components/dependencies.sh
@@ -1,27 +1,33 @@
 function pod_install() {
   comp_init 'dependencies'
-  check_deps 'pod'
-  msg 'Installing CocoaPods'
-  if [[ -z $clean ]]; then
-    args="--no-repo-update $args"
+  if [[ -f Podfile ]]; then
+    check_deps 'pod'
+    msg 'Installing CocoaPods'
+    if [[ -z $clean ]]; then
+      args="--no-repo-update $args"
+    fi
+    bundle exec pod install $args $@
   fi
-  bundle exec pod install $args $@
   comp_deinit
 }
 
 function carthage_bootstrap() {
   comp_init 'dependencies'
-  check_deps 'carthage'
-  msg 'Installing Carthage packages'
-  run_carthage bootstrap
+  if [[ -f Cartfile ]]; then
+    check_deps 'carthage'
+    msg 'Installing Carthage packages'
+    run_carthage bootstrap
+  fi
   comp_deinit
 }
 
 function carthage_update() {
   comp_init 'dependencies'
-  check_deps 'carthage'
-  msg 'Installing Carthage packages'
-  run_carthage update
+  if [[ -f Cartfile ]]; then
+    check_deps 'carthage'
+    msg 'Installing Carthage packages'
+    run_carthage update
+  fi
   comp_deinit
 }
